+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/framework/ExpenseClaimController.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 
 import android.content.Context;
-import ca.ualberta.cs.shinyexpensetracker.models.AbstractExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseItem;
@@ -103,7 +102,7 @@ public class ExpenseClaimController {
 	 * Returns the model
 	 * @return
 	 */
-	public AbstractExpenseClaimList getExpenseClaimList() {
+	public ExpenseClaimList getExpenseClaimList() {
 		return claimList;
 	}
